fix(website): run .md mirror as post-build step, not Docusaurus plugin
Docusaurus 3.x runs plugins' postBuild hooks concurrently (Promise.all),
so the inline mirror plugin raced against docusaurus-plugin-llms and
walked the build directory before the llms plugin had written the
individual .md files. That left every mirrored URL (e.g.
`/reference/schedule.md`, `/reference/di.md`) 404 on the preview.

Move the mirror to `website/scripts/llms-md-mirror.js`, invoked from
`package.json`'s `build` script with `&&` after `docusaurus build`
completes. At that point every plugin's postBuild has resolved, so
the walk sees the final tree. Update the rationale in
`docs/contributing-to-documentation.md` to reflect the new location
and explain why the plugin approach didn't work.
