+++ website/scripts/llms-md-mirror.js
@@ -1,0 +1,54 @@
+#!/usr/bin/env node
+// Mirror .md files generated by docusaurus-plugin-llms to the canonical URL paths
+// a visitor would naively construct by appending `.md` to any rendered URL.
+//
+// See docs/contributing-to-documentation.md ("AI-friendly Markdown Variants")
+// for the rationale. Runs after `docusaurus build` so it sees every .md that
+// docusaurus-plugin-llms wrote — Docusaurus' postBuild hooks run concurrently,
+// so doing this as another plugin would race against the llms plugin.
+
+const fs = require('fs').promises;
+const path = require('path');
+
+async function mirror(outDir, dir) {
+  let created = 0;
+  const entries = await fs.readdir(dir, { withFileTypes: true });
+  for (const entry of entries) {
+    const full = path.join(dir, entry.name);
+    if (entry.isDirectory()) {
+      created += await mirror(outDir, full);
+      continue;
+    }
+    if (!entry.isFile() || !entry.name.endsWith('.md')) continue;
+    const parent = path.dirname(full);
+    if (parent === outDir) continue;
+    const base = entry.name.replace(/\.md$/, '');
+    if (base !== 'index' && base !== path.basename(parent)) continue;
+    const target = `${parent}.md`;
+    try {
+      await fs.access(target);
+    } catch {
+      await fs.copyFile(full, target);
+      created += 1;
+    }
+  }
+  return created;
+}
+
+async function main() {
+  const outDir = path.resolve(process.argv[2] || 'build');
+  try {
+    await fs.access(outDir);
+  } catch {
+    console.error(`llms-md-mirror: build directory not found: ${outDir}`);
+    process.exit(1);
+  }
+  const created = await mirror(outDir, outDir);
+  console.log(`llms-md-mirror: created ${created} mirrored .md file(s) in ${outDir}`);
+}
+
+main().catch((err) => {
+  console.error('llms-md-mirror: failed');
+  console.error(err);
+  process.exit(1);
+});
